Add appropriate linking to `libssh`
I'm not sure why but in `ssh2-rs` this does not seem to be needed,
perhaps there's a difference here related to the dynamic v. static
linking.

Rust by Example states that the "link" is needed, this makes sense, how
else would the compiler know where to look for the functions?

https://rustbyexample.com/std_misc/ffi.html